+++ db/migrate/20230315141722_create_debits.rb
@@ -3,7 +3,7 @@ class CreateDebits < ActiveRecord::Migration[7.0]
     create_table :debits do |t|
       t.string :name
       t.text :description
-      t.float :balance
+      t.decimal :balance
       t.references :user, null: false, foreign_key: true
 
       t.timestamps
